+++ core/user/repository.go
@@ -20,5 +20,6 @@ type Repository interface {
 	) (*pagination.Page[User], error)
 	IsEnabledByID(ctx context.Context, id uuid.UUID) (bool, error)
 	Count(ctx context.Context) (int, error)
+	TryCreateInitialUser(ctx context.Context, user *User) (bool, error)
 	TryUpdateLastUsedTOTPCode(ctx context.Context, id uuid.UUID, code string) (bool, error)
 }